+++ todolist/todolist_app/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
     path("todolist/<int:todo_id>/", views.todolist_update, name = "todolist_update"),
     path("item/create/", views.item_create, name = "item_create"),
     path("item/update/", views.item_update, name = "item_update"),
+    path("item/delete/<int:item_id>/", views.item_delete, name = "item_delete"),
 ]